Refactor: getOGImage 함수 분리
